Rewrite the end section; clarify enforcement
Incorporate suggestions from dove and others to clarify how
enforcement will work.

Hopefully this section now communicates that this will be largely
community enforced, but will not preclude anyone who has been hurt by
a "violation" of these rules from coming to the EC for assistance.
The goal is to make it clear that there are not specific punishments
for each instance of "well-actually" or "Sorry but not sorry", but
simultaneously not in any way deny the EC the power to take action
against any SIPB member or propsective who is making the office an
unwelcoming (or even hostile) environment.

Third parties are encouraged to talk to the recipient first, but
should still feel empowered to speak up if necessary.  (e.g. If $X
says something sexist to $Y, and $Y doesn't react, but $Z also was
upset by the sexist comment, $Z should feel free to speak out.)  The
last sentence in the penultimate paragraph is deliberately the same as
in the paragraph above it.