Rewrite last section
Update the "What happens if" section to make it clear that public
callouts in electronic media will not end well, and that public
callouts in person will fare better if you focus on the statement
or action, not the person making it.  The goal here is to convey
the difference between, e.g. "I found that statement very sexist"
vs "You are being sexist".  The former will result in far less
defensiveness.